+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 ### Changed
 - Updated default sandbox version to `97c0410de519ecaca369aaee26f0ca5eb9e7de06` commit of nearcore to include 1.26 protocol changes https://github.com/near/workspaces-rs/pull/134
 
+- Exposed `CallExecutionDetails::raw_bytes` API: https://github.com/near/workspaces-rs/pull/133
+
 ## [0.2.1] - 2022-04-12
 
 ### Added